% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_rownumber.R
\name{fill_missing_rownumber}
\alias{fill_missing_rownumber}
\title{Fill missing rownumber}
\usage{
fill_missing_rownumber(x)
}
\arguments{
\item{x}{een vector met integers}
}
\value{
x, met de NA's opgevuld indien mogelijk
}
\description{
Een functie om missende waarden van een telvariabele op te vullen. De ontbrekende
missende waarden worden opgevuld door door te tellen vanaf de laatst bekende
waarde.
c(NA,4,NA,NA) wordt bijvoorbeeld c(NA,4,5,6)
}
\examples{
fill_missing_rownumber(c(NA,4,NA,NA))
}
\concept{missing data functies}
\concept{vector berekeningen}
