% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing.R
\name{fill_with_agg_by_group}
\alias{fill_with_agg_by_group}
\title{Fill with aggregate by group}
\usage{
fill_with_agg_by_group(df, group, col, overwrite_col = F, statistic = mean)
}
\arguments{
\item{df}{tibble to use}

\item{group}{string or vector of strings: columns to group by}

\item{col}{string: column to impute}

\item{overwrite_col}{boolean: whether to overwrite column. If FALSE, a new column with suffix _imputed will be created}

\item{statistic}{function: summary statistic to use (mean, median, min etc.). For now requires a function with na.rm argument}
}
\value{
a tibble with filled column
}
\description{
Function to calculate a summary statistic (mean, median, vvconverter::mode, min, max etc.) by group
and use it to fill missing values. Note: this takes and produces a tibble rather than a vector.
}
