% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_impute_median.R
\name{NA_impute_median}
\alias{NA_impute_median}
\title{NA impute median}
\usage{
NA_impute_median(data, var, year = 2014)
}
\arguments{
\item{data}{The data frame.}

\item{var}{The variable used to create new variables.}

\item{year}{Year used for the median for imputation.}
}
\value{
New data frame in which missing values are filled.
}
\description{
Is a specialised function which takes a variable and turns it into two new variables
to be used in a prediction model.
\enumerate{
\item the variable for which missing values are imputed by the median for the given year.
\item an indicator when the variable is missing
}
}
