% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_with_agg.R
\name{fill_col_with_agg_by_group}
\alias{fill_col_with_agg_by_group}
\title{Fill column with aggregate by group}
\usage{
fill_col_with_agg_by_group(df, group, col, statistic)
}
\arguments{
\item{df}{tibble to use}

\item{group}{string or vector of strings: columns to group by}

\item{col}{string: column to impute}

\item{statistic}{function: summary statistic to use (mean, median, min etc.). For now requires a function with na.rm argument}
}
\value{
a filled vector
}
\description{
Calculate a summary statistic (mean, median, vvconverter::mode, min, max etc.) by group
and use it to fill missing values in a column. Primarily for use in fill_with_agg_by_group().
}
