% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_rownumber.R
\name{fill_missing_rownumber}
\alias{fill_missing_rownumber}
\title{Fill missing rownumber}
\usage{
fill_missing_rownumber(x)
}
\arguments{
\item{x}{Integer vector.}
}
\value{
Integer vector with filled values.
}
\description{
Impute missing values of a count variable. Imputation is done by counting from
the last known value. Example: c(NA,4,NA,NA) then becomes c(NA,4,NA,NA).
}
\examples{
fill_missing_rownumber(c(NA,4,NA,NA))
}
\concept{missing data functions}
\concept{vector calculations}
