% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_impute_median.R
\name{NA_impute_median}
\alias{NA_impute_median}
\title{NA impute median}
\usage{
NA_impute_median(data, var, jaar = 2014)
}
\arguments{
\item{data}{De data frame.}

\item{var}{De variabele waarmee nieuwe variabelen worden aangemaakt.}

\item{jaar}{Jaartal waarmee missende waarden worden opgevuld door de mediaan.}
}
\value{
Nieuwe data frame waarbij de missende waarden zijn opgevuld.
}
\description{
Is een specifieke functie die op basis van 1 variabele, 2 nieuwe variabelen aanmaakt
voor gebruik in een voorspelmodel:
\enumerate{
\item de variabele, waarbij de missende waarden opgevuld zijn door de mediaan van
het opgegeven jaartal.
\item een indicator voor wanneer de variabele missing is.
}
}
