% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_impute_median.R
\name{na_impute_median}
\alias{na_impute_median}
\title{NA impute median}
\usage{
na_impute_median(data, var, year = 2014, year_column)
}
\arguments{
\item{data}{The data frame.}

\item{var}{The variable used to create new variables.}

\item{year}{Year used for the median for imputation.}

\item{year_column}{Column with year to use median on.}
}
\value{
New data frame in which missing values are filled.
}
\description{
Is a specialized function which takes a variable and turns it into two new variables
to be used in a prediction model.
\enumerate{
\item the variable for which missing values are imputed by the median for the given year.
\item an indicator when the variable is missing
}
}
